Apply improved types & architecture to Hedera
