+++ libs/ledger-live-common/src/families/hedera/bridge/js.ts
@@ -1,12 +1,13 @@
 import type { AccountBridge, CurrencyBridge } from "@ledgerhq/types-live";
 import type { Transaction } from "../types";
-import getTransactionStatus from "../js-getTransactionStatus";
-import estimateMaxSpendable from "../js-estimateMaxSpendable";
-import signOperation from "../js-signOperation";
-import broadcast from "../js-broadcast";
-import { scanAccounts, sync } from "../js-synchronisation";
-import receive from "../js-receive";
-import { createTransaction, prepareTransaction } from "../js-transaction";
+import { getTransactionStatus } from "../getTransactionStatus";
+import { estimateMaxSpendable } from "../estimateMaxSpendable";
+import { signOperation } from "../signOperation";
+import { broadcast } from "../broadcast";
+import { scanAccounts, sync } from "../synchronisation";
+import { receive } from "../receive";
+import { prepareTransaction } from "../prepareTransaction";
+import { createTransaction } from "../createTransaction";
 import { defaultUpdateTransaction } from "@ledgerhq/coin-framework/bridge/jsHelpers";
 
 const currencyBridge: CurrencyBridge = {
