+++ libs/ledger-live-common/src/families/hedera/prepareTransaction.ts
@@ -1,21 +1,6 @@
-import BigNumber from "bignumber.js";
-import type { Account } from "@ledgerhq/types-live";
+import type { AccountBridge } from "@ledgerhq/types-live";
 import type { Transaction } from "./types";
 import { calculateAmount } from "./utils";
-
-/**
- * Creates an empty transaction.
- *
- * @returns {Transaction}
- */
-export function createTransaction(_account: Account): Transaction {
-  return {
-    family: "hedera",
-    amount: new BigNumber(0),
-    recipient: "",
-    useAllAmount: false,
-  };
-}
 
 /**
  * Gather any more neccessary information for a transaction,
@@ -26,15 +11,14 @@ export function createTransaction(_account: Account): Transaction {
  *
  * @returns  {Transaction}
  */
-export async function prepareTransaction(
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  account: Account,
-  transaction: Transaction,
-): Promise<Transaction> {
+export const prepareTransaction: AccountBridge<Transaction>["prepareTransaction"] = async (
+  account,
+  transaction,
+) => {
   // explicitly calculate transaction amount to account for `useAllAmount` flag (send max flow)
   // i.e. if `useAllAmount` has been toggled to true, this is where it will update the transaction to reflect that action
   const { amount } = await calculateAmount({ account, transaction });
   transaction.amount = amount;
 
   return transaction;
-}
+};
