+++ libs/ledger-live-common/src/families/hedera/synchronisation.ts
