+++ libs/ledger-live-common/src/families/hedera/tests/js-transaction.test.ts
@@ -1,8 +1,9 @@
 import BigNumber from "bignumber.js";
 import type { Account } from "@ledgerhq/types-live";
-import type { Transaction } from "../../../families/hedera/types";
 import { defaultUpdateTransaction as updateTransaction } from "@ledgerhq/coin-framework/bridge/jsHelpers";
-import { createTransaction, prepareTransaction } from "../js-transaction";
+import type { Transaction } from "../../../families/hedera/types";
+import { createTransaction } from "../createTransaction";
+import { prepareTransaction } from "../prepareTransaction";
 
 const account: Account = {
   type: "Account",
